LayoutTests/media/modern-media-controls/layout-node/x-expected.txt

Testing the LayoutNode.x property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Checking default value
PASS node.x is 0
PASS node.element.style.left is ""

node.x = 10
PASS node.x is 10
PASS node.element.style.left is ""

node.x = 20

Layout was performed
PASS node.x is 20
PASS node.element.style.left is "20px"

node.x = 20
PASS scheduler.hasScheduledLayoutCallbacks is false

PASS successfullyParsed is true

TEST COMPLETE
